Create new page: add_expense_page_content and cubit
