+++ lib/app/features/home/add_expense/cubit/add_expense_state.dart
@@ -1,0 +1,11 @@
+part of 'add_expense_cubit.dart';
+
+class AddExpenseState {
+  final String errorMessage;
+  final bool save;
+
+  AddExpenseState({
+    this.errorMessage = '',
+    this.save = false,
+  });
+}
