+++ lib/app/features/home/home_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:skarbonka_v2/app/cubit/root_cubit.dart';
-import 'package:skarbonka_v2/app/features/home/expense_list/expense_list_page_content.dart';
+import 'package:skarbonka_v2/app/features/home/expense_list/page/expense_list_page_content.dart';
 import 'package:skarbonka_v2/app/features/home/monthly_report/monthly_report_page_content.dart';
-import 'package:skarbonka_v2/app/features/home/my_account/my_account_page_content.dart';
+import 'package:skarbonka_v2/app/features/home/my_account/page/my_account_page_content.dart';
 
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:convex_bottom_bar/convex_bottom_bar.dart';
